HITO 2 - Modelo de Compras con uso de BigDecimal

+++ backend/src/main/java/com/example/backend/dto/request/CompraRequestDTO.java
@@ -1,10 +1,11 @@
 package com.example.backend.dto.request;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record CompraRequestDTO(
     Long id,                  // para ediciones, se debe omitir en creaciones
     String proveedorNombre,   //se debe validar que exista un proveedor con ese nombre
-    Double totalCompra,       // validar que sea igual a la suma de los subtotales de los detalles
+    BigDecimal totalCompra,       // validar que sea igual a la suma de los subtotales de los detalles
     List<DetalleCompraRequestDTO> detalles
 ) {}
