+++ backend/src/main/java/com/example/backend/dto/request/DetalleCompraRequestDTO.java
@@ -1,8 +1,9 @@
 package com.example.backend.dto.request;
 
+import java.math.BigDecimal;
 
 public record DetalleCompraRequestDTO(
     ProductoIdDTO producto,   // Ataja el JSON anidado: { "id": X }
     Integer cantidad,
-    Double precioCompra      
+    BigDecimal precioCompra      
 ) {}
