+++ backend/src/main/java/com/example/backend/dto/response/DetalleCompraResponseDTO.java
@@ -1,5 +1,7 @@
 package com.example.backend.dto.response;
 import lombok.*;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -10,7 +12,7 @@ public class DetalleCompraResponseDTO {
     private Long id;
     private List<ProductoResponseCompraDTO> producto;
     private Integer cantidad;
-    private Double precioUnitario;
-    private Double subtotal;
+    private BigDecimal precioUnitario;
+    private BigDecimal subtotal;
     private String nombreprovedor;
 }
